Human Microbiome Project WGS .

Whole Genome Sequencing data obtained from the Human Microbiome Project production phase I.

Usage

hmp_wgs

Format

hmp_wgs

Data frame that contains.

PID

Module Table ID that identifies organisms that contain complete KEGG modules

Abundance

Abundance data

...

Source

https://portal.hmpdacc.org/files/1670203039de370df9a35a043735e964

Details

"The overall mission of the HMP is to generate resources to facilitate characterization of the human microbiota to further our understanding of how the microbiome impacts human health and disease. We are currently in the second phase of the HMP. In this phase, the iHMP will create integrated longitudinal datasets from both the microbiome and host from three different cohort studies of microbiome-associated conditions using multiple ‘omics technologies. Each of these study groups has engaged in providing new computational tools and integrative molecular perspectives on microbial activity during dysbiosis. As a result of creating these multi’omic data resources, the iHMP has opened up new opportunities for data integration in the human microbiome.". https://www.hmpdacc.org/ihmp/