Predict Rank Abundance Distribution

Uses the parameters calculated by BC_param to model a DGB distribution (Mansilla et al. (2007) doi:10.1016/j.joi.2007.01.001 ) from the rank information in the data frame.

Usage

BC_model(df_abundance = NULL, column = NULL, BC_param_object = NULL, ...)

Arguments

df_abundance

A data frame that contains abundance data.

column

Either a string with the name of the column or the number of the column that stores the abundances in the data frame.

BC_param_object

Optional. A previous object generated by BC_param.

...

passes arguments to BC_param.

Value

A list with the following elements: The input data frame with added processed ranking data, model data and confidence interval data, the adjusted parameters, the confidence interval of the parameters, the linear model, a summary of the model and a generated function for use with raw numeric data.

Examples

BC_model(Citation_ISICatalogued, column = 2, show_stats = F,rank_threshold=1)
#> [[1]]
#>     pre_numerator pre_denominator          lwr predicted_values          upr
#> 1             486               1 2.797814e+06     3.399435e+06 4.130425e+06
#> 2             485               2 6.177875e+05     7.298173e+05 8.621625e+05
#> 3             484               3 2.552310e+05     2.966446e+05 3.447779e+05
#> 4             483               4 1.362779e+05     1.565871e+05 1.799229e+05
#> 5             482               5 8.374506e+04     9.538180e+04 1.086355e+05
#> 6             481               6 5.624755e+04     6.360792e+04 7.193145e+04
#> 7             480               7 4.016922e+04     4.515464e+04 5.075879e+04
#> 8             479               8 3.000447e+04     3.355537e+04 3.752650e+04
#> 9             478               9 2.319392e+04     2.582239e+04 2.874875e+04
#> 10            477              10 1.842096e+04     2.042683e+04 2.265113e+04
#> 11            476              11 1.495396e+04     1.652303e+04 1.825673e+04
#> 12            475              12 1.236089e+04     1.361365e+04 1.499338e+04
#> 13            474              13 1.037371e+04     1.139132e+04 1.250876e+04
#> 14            473              14 8.819247e+03     9.658102e+03 1.057675e+04
#> 15            472              15 7.581703e+03     8.282052e+03 9.047095e+03
#> 16            471              16 6.581346e+03     7.172587e+03 7.816943e+03
#> 17            470              17 5.761879e+03     6.265917e+03 6.814047e+03
#> 18            469              18 5.082666e+03     5.516111e+03 5.986519e+03
#> 19            468              19 4.513798e+03     4.889444e+03 5.296351e+03
#> 20            467              20 4.032885e+03     4.360721e+03 4.715207e+03
#> 21            466              21 3.622901e+03     3.910828e+03 4.221637e+03
#> 22            465              22 3.270722e+03     3.525056e+03 3.799168e+03
#> 23            464              23 2.966103e+03     3.191948e+03 3.434989e+03
#> 24            463              24 2.700958e+03     2.902473e+03 3.119024e+03
#> 25            462              25 2.468835e+03     2.649442e+03 2.843262e+03
#> 26            461              26 2.264541e+03     2.427075e+03 2.601275e+03
#> 27            460              27 2.083855e+03     2.230682e+03 2.387855e+03
#> 28            459              28 1.923324e+03     2.056433e+03 2.198753e+03
#> 29            458              29 1.780097e+03     1.901168e+03 2.030473e+03
#> 30            457              30 1.651808e+03     1.762270e+03 1.880119e+03
#> 31            456              31 1.536476e+03     1.637551e+03 1.745276e+03
#> 32            455              32 1.432439e+03     1.525177e+03 1.623918e+03
#> 33            454              33 1.338288e+03     1.423594e+03 1.514338e+03
#> 34            453              34 1.252827e+03     1.331487e+03 1.415086e+03
#> 35            451              36 1.104029e+03     1.171362e+03 1.242802e+03
#> 36            452              35 1.175033e+03     1.247731e+03 1.324927e+03
#> 37            450              37 1.039059e+03     1.101552e+03 1.167802e+03
#> 38            449              38 9.794682e+02     1.037581e+03 1.099141e+03
#> 39            448              39 9.246866e+02     9.788268e+02 1.036137e+03
#> 40            447              40 8.742184e+02     9.247470e+02 9.781961e+02
#> 41            445              42 7.845368e+02     8.287690e+02 8.754949e+02
#> 42            446              41 8.276290e+02     8.748666e+02 9.248003e+02
#> 43            443              44 7.075385e+02     7.464986e+02 7.876039e+02
#> 44            444              43 7.446057e+02     7.860875e+02 8.298803e+02
#> 45            442              45 6.730721e+02     7.097155e+02 7.483538e+02
#> 46            440              47 6.110310e+02     6.435768e+02 6.778560e+02
#> 47            441              46 6.409724e+02     6.754837e+02 7.118532e+02
#> 48            439              48 5.830616e+02     6.137923e+02 6.461428e+02
#> 49            438              49 5.568976e+02     5.859496e+02 6.165171e+02
#> 50            435              52 4.878150e+02     5.125290e+02 5.384951e+02
#> 51            436              51 5.094022e+02     5.354565e+02 5.628433e+02
#> 52            437              50 5.323894e+02     5.598862e+02 5.888033e+02
#> 53            434              53 4.675180e+02     4.909853e+02 5.156305e+02
#> 54            433              54 4.484122e+02     4.707181e+02 4.941336e+02
#> 55            432              55 4.304077e+02     4.516305e+02 4.738996e+02
#> 56            431              56 4.134229e+02     4.336341e+02 4.548334e+02
#> 57            430              57 3.973834e+02     4.166489e+02 4.368484e+02
#> 58            429              58 3.822216e+02     4.006019e+02 4.198661e+02
#> 59            427              60 3.542891e+02     3.710619e+02 3.886288e+02
#> 60            428              59 3.678757e+02     3.854265e+02 4.038148e+02
#> 61            426              61 3.414101e+02     3.574523e+02 3.742483e+02
#> 62            424              63 3.175892e+02     3.322984e+02 3.476888e+02
#> 63            423              64 3.065635e+02     3.206641e+02 3.354133e+02
#> 64            425              62 3.291914e+02     3.445467e+02 3.606184e+02
#> 65            422              65 2.960775e+02     3.096044e+02 3.237492e+02
#> 66            419              68 2.675297e+02     2.795210e+02 2.920497e+02
#> 67            421              66 2.860970e+02     2.990826e+02 3.126576e+02
#> 68            420              67 2.765908e+02     2.890652e+02 3.021022e+02
#> 69            418              69 2.588869e+02     2.704211e+02 2.824693e+02
#> 70            416              71 2.427584e+02     2.534501e+02 2.646128e+02
#> 71            414              73 2.280270e+02     2.379613e+02 2.483284e+02
#> 72            417              70 2.506375e+02     2.617391e+02 2.733325e+02
#> 73            415              72 2.352282e+02     2.455313e+02 2.562856e+02
#> 74            412              75 2.145386e+02     2.237900e+02 2.334403e+02
#> 75            411              76 2.082182e+02     2.171531e+02 2.264715e+02
#> 76            410              77 2.021598e+02     2.107937e+02 2.197963e+02
#> 77            413              74 2.211362e+02     2.307204e+02 2.407199e+02
#> 78            409              78 1.963495e+02     2.046967e+02 2.133988e+02
#> 79            406              81 1.802802e+02     1.878454e+02 1.957281e+02
#> 80            408              79 1.907742e+02     1.988482e+02 2.072640e+02
#> 81            405              82 1.753393e+02     1.826674e+02 1.903017e+02
#> 82            407              80 1.854216e+02     1.932352e+02 2.013781e+02
#> 83            403              84 1.660193e+02     1.729044e+02 1.800750e+02
#> 84            404              83 1.705889e+02     1.776904e+02 1.850876e+02
#> 85            402              85 1.616219e+02     1.682999e+02 1.752540e+02
#> 86            398              89 1.455925e+02     1.515273e+02 1.577041e+02
#> 87            399              88 1.493806e+02     1.554894e+02 1.618481e+02
#> 88            401              86 1.573881e+02     1.638681e+02 1.706150e+02
#> 89            397              90 1.419393e+02     1.477073e+02 1.537097e+02
#> 90            400              87 1.533101e+02     1.596006e+02 1.661491e+02
#> 91            391              96 1.225289e+02     1.274268e+02 1.325206e+02
#> 92            395              92 1.350131e+02     1.404675e+02 1.461422e+02
#> 93            394              93 1.317288e+02     1.370356e+02 1.425563e+02
#> 94            390              97 1.196643e+02     1.244363e+02 1.293986e+02
#> 95            387             100 1.116183e+02     1.160398e+02 1.206365e+02
#> 96            396              91 1.384148e+02     1.440228e+02 1.498579e+02
#> 97            393              94 1.285565e+02     1.337217e+02 1.390943e+02
#> 98            388              99 1.142129e+02     1.187469e+02 1.234608e+02
#> 99            392              95 1.254915e+02     1.305204e+02 1.357509e+02
#> 100           389              98 1.168937e+02     1.215444e+02 1.263801e+02
#> 101           386             101 1.091062e+02     1.134194e+02 1.179030e+02
#> 102           385             102 1.066734e+02     1.108820e+02 1.152567e+02
#> 103           380             107 9.558942e+01     9.932803e+01 1.032129e+02
#> 104           379             108 9.356929e+01     9.722330e+01 1.010200e+02
#> 105           384             103 1.043165e+02     1.084243e+02 1.126940e+02
#> 106           383             104 1.020325e+02     1.060431e+02 1.102114e+02
#> 107           382             105 9.981844e+01     1.037352e+02 1.078057e+02
#> 108           381             106 9.767164e+01     1.014978e+02 1.054738e+02
#> 109           378             109 9.160884e+01     9.518107e+01 9.889260e+01
#> 110           375             112 8.606341e+01     8.940602e+01 9.287846e+01
#> 111           377             110 8.970581e+01     9.319896e+01 9.682813e+01
#> 112           361             126 6.556077e+01     6.807622e+01 7.068817e+01
#> 113           370             117 7.782191e+01     8.082791e+01 8.395003e+01
#> 114           374             113 8.432000e+01     8.759094e+01 9.098877e+01
#> 115           365             122 7.064833e+01     7.336586e+01 7.618792e+01
#> 116           371             116 7.937854e+01     8.244770e+01 8.563553e+01
#> 117           373             114 8.262591e+01     8.582745e+01 8.915303e+01
#> 118           367             120 7.340154e+01     7.622929e+01 7.916598e+01
#> 119           366             121 7.200646e+01     7.477829e+01 7.765681e+01
#> 120           368             119 7.483488e+01     7.772026e+01 8.071689e+01
#> 121           369             118 7.630785e+01     7.925263e+01 8.231104e+01
#> 122           372             115 8.097933e+01     8.411363e+01 8.736925e+01
#> 123           376             111 8.785803e+01     9.127467e+01 9.482418e+01
#> 124           363             124 6.803790e+01     7.065149e+01 7.336547e+01
#> 125           360             127 6.436944e+01     6.683786e+01 6.940093e+01
#> 126           359             128 6.320822e+01     6.563091e+01 6.814645e+01
#> 127           362             125 6.678324e+01     6.934705e+01 7.200929e+01
#> 128           364             123 6.932588e+01     7.199068e+01 7.475792e+01
#> 129           353             134 5.682087e+01     5.899384e+01 6.124990e+01
#> 130           355             132 5.884539e+01     6.109717e+01 6.343512e+01
#> 131           357             130 6.097223e+01     6.330714e+01 6.573147e+01
#> 132           358             129 6.207613e+01     6.445433e+01 6.692364e+01
#> 133           350             137 5.396231e+01     5.602448e+01 5.816546e+01
#> 134           348             139 5.216704e+01     5.415993e+01 5.622895e+01
#> 135           344             143 4.881836e+01     5.068258e+01 5.261799e+01
#> 136           352             135 5.584500e+01     5.798008e+01 6.019678e+01
#> 137           354             133 5.782074e+01     6.003260e+01 6.232906e+01
#> 138           345             142 4.962680e+01     5.152201e+01 5.348961e+01
#> 139           347             140 5.130055e+01     5.326008e+01 5.529445e+01
#> 140           341             146 4.650031e+01     4.827588e+01 5.011924e+01
#> 141           351             136 5.489239e+01     5.699054e+01 5.916890e+01
#> 142           349             138 5.305408e+01     5.508117e+01 5.718572e+01
#> 143           327             160 3.748226e+01     3.891568e+01 4.040391e+01
#> 144           333             154 4.102050e+01     4.258770e+01 4.421478e+01
#> 145           331             156 3.979155e+01     4.131223e+01 4.289102e+01
#> 146           342             145 4.725564e+01     4.906006e+01 5.093338e+01
#> 147           330             157 3.919615e+01     4.069431e+01 4.224974e+01
#> 148           329             158 3.861302e+01     4.008914e+01 4.162169e+01
#> 149           340             147 4.576167e+01     4.750905e+01 4.932316e+01
#> 150           343             144 4.802816e+01     4.986212e+01 5.176611e+01
#> 151           346             141 5.045400e+01     5.238098e+01 5.438155e+01
#> 152           338             149 4.433256e+01     4.602551e+01 4.778311e+01
#> 153           335             152 4.230270e+01     4.391852e+01 4.559606e+01
#> 154           336             151 4.296471e+01     4.460566e+01 4.630929e+01
#> 155           339             148 4.503924e+01     4.675909e+01 4.854462e+01
#> 156           326             161 3.693400e+01     3.834674e+01 3.981351e+01
#> 157           311             176 2.988780e+01     3.103531e+01 3.222688e+01
#> 158           337             150 4.364120e+01     4.530785e+01 4.703816e+01
#> 159           322             165 3.484839e+01     3.618252e+01 3.756772e+01
#> 160           316             171 3.201333e+01     3.324076e+01 3.451525e+01
#> 161           334             153 4.165476e+01     4.324600e+01 4.489803e+01
#> 162           305             182 2.758469e+01     2.864559e+01 2.974729e+01
#> 163           315             172 3.157198e+01     3.278281e+01 3.404008e+01
#> 164           325             162 3.639677e+01     3.778924e+01 3.923499e+01
#> 165           317             170 3.246321e+01     3.370756e+01 3.499961e+01
#> 166           324             163 3.587028e+01     3.724290e+01 3.866805e+01
#> 167           304             183 2.722467e+01     2.827203e+01 2.935968e+01
#> 168           321             166 3.435248e+01     3.566793e+01 3.703375e+01
#> 169           323             164 3.535424e+01     3.670742e+01 3.811239e+01
#> 170           309             178 2.909176e+01     3.020934e+01 3.136985e+01
#> 171           332             155 4.039955e+01     4.194324e+01 4.354591e+01
#> 172           320             167 3.386624e+01     3.516339e+01 3.651022e+01
#> 173           314             173 3.113895e+01     3.233350e+01 3.357387e+01
#> 174           298             189 2.519384e+01     2.616474e+01 2.717306e+01
#> 175           308             179 2.870462e+01     2.980764e+01 3.095305e+01
#> 176           310             177 2.948610e+01     3.061850e+01 3.179440e+01
#> 177           294             193 2.395323e+01     2.487736e+01 2.583714e+01
#> 178           312             175 3.029705e+01     3.145995e+01 3.266748e+01
#> 179           303             184 2.687105e+01     2.790510e+01 2.897894e+01
#> 180           328             159 3.804183e+01     3.949637e+01 4.100653e+01
#> 181           307             180 2.832450e+01     2.941323e+01 3.054380e+01
#> 182           318             169 3.292184e+01     3.418345e+01 3.549340e+01
#> 183           272             215 1.842032e+01     1.913468e+01 1.987676e+01
#> 184           301             186 2.618240e+01     2.719053e+01 2.823748e+01
#> 185           296             191 2.456288e+01     2.551000e+01 2.649364e+01
#> 186           297             190 2.487564e+01     2.583455e+01 2.683042e+01
#> 187           319             168 3.338944e+01     3.466864e+01 3.599685e+01
#> 188           291             196 2.307671e+01     2.396775e+01 2.489319e+01
#> 189           277             210 1.951285e+01     2.026884e+01 2.105412e+01
#> 190           313             174 3.071404e+01     3.189261e+01 3.311641e+01
#> 191           288             199 2.224314e+01     2.310268e+01 2.399543e+01
#> 192           299             188 2.551762e+01     2.650071e+01 2.752169e+01
#> 193           356             131 5.989560e+01     6.218839e+01 6.456895e+01
#> 194           302             185 2.652367e+01     2.754465e+01 2.860493e+01
#> 195           282             205 2.069432e+01     2.149519e+01 2.232705e+01
#> 196           295             192 2.425545e+01     2.519098e+01 2.616259e+01
#> 197           287             200 2.197436e+01     2.282373e+01 2.370593e+01
#> 198           306             181 2.795125e+01     2.902593e+01 3.014194e+01
#> 199           276             211 1.928752e+01     2.003493e+01 2.081131e+01
#> 200           290             197 2.279422e+01     2.367458e+01 2.458895e+01
#> 201           286             201 2.170996e+01     2.254932e+01 2.342113e+01
#> 202           270             217 1.800626e+01     1.870482e+01 1.943048e+01
#> 203           293             194 2.365611e+01     2.456902e+01 2.551717e+01
#> 204           281             206 2.045046e+01     2.124208e+01 2.206434e+01
#> 205           275             212 1.906567e+01     1.980463e+01 2.057224e+01
#> 206           285             202 2.144984e+01     2.227935e+01 2.314094e+01
#> 207           300             187 2.584709e+01     2.684260e+01 2.787645e+01
#> 208           292             195 2.336397e+01     2.426586e+01 2.520256e+01
#> 209           280             207 2.021047e+01     2.099297e+01 2.180577e+01
#> 210           284             203 2.119392e+01     2.201373e+01 2.286526e+01
#> 211           289             198 2.251640e+01     2.338626e+01 2.428973e+01
#> 212           279             208 1.997425e+01     2.074778e+01 2.155127e+01
#> 213           265             222 1.702414e+01     1.768512e+01 1.837176e+01
#> 214           283             204 2.094211e+01     2.175238e+01 2.259399e+01
#> 215           242             245 1.331730e+01     1.383516e+01 1.437316e+01
#> 216           264             223 1.683634e+01     1.749012e+01 1.816929e+01
#> 217           269             218 1.780390e+01     1.849472e+01 1.921235e+01
#> 218           268             219 1.760456e+01     1.828776e+01 1.899748e+01
#> 219           241             246 1.318138e+01     1.369395e+01 1.422645e+01
#> 220           240             247 1.304726e+01     1.355461e+01 1.408168e+01
#> 221           239             248 1.291492e+01     1.341710e+01 1.393881e+01
#> 222           252             235 1.478302e+01     1.535774e+01 1.595480e+01
#> 223           263             224 1.665129e+01     1.729797e+01 1.796977e+01
#> 224           267             220 1.740819e+01     1.808388e+01 1.878580e+01
#> 225           262             225 1.646894e+01     1.710862e+01 1.777315e+01
#> 226           261             226 1.628923e+01     1.692201e+01 1.757937e+01
#> 227           260             227 1.611212e+01     1.673809e+01 1.738837e+01
#> 228           278             209 1.974174e+01     2.050643e+01 2.130075e+01
#> 229           229             258 1.168247e+01     1.213644e+01 1.260805e+01
#> 230           274             213 1.884723e+01     1.957787e+01 2.033684e+01
#> 231           249             238 1.432198e+01     1.487886e+01 1.545739e+01
#> 232           182             305 7.533797e+00     7.823545e+00 8.124436e+00
#> 233           238             249 1.278431e+01     1.328140e+01 1.379782e+01
#> 234           271             216 1.821171e+01     1.891812e+01 1.965192e+01
#> 235           228             259 1.156773e+01     1.201719e+01 1.248412e+01
#> 236           258             229 1.576548e+01     1.637811e+01 1.701455e+01
#> 237           246             241 1.387965e+01     1.441938e+01 1.498009e+01
#> 238           237             250 1.265542e+01     1.314747e+01 1.365866e+01
#> 239           227             260 1.145443e+01     1.189944e+01 1.236174e+01
#> 240           273             214 1.863213e+01     1.935458e+01 2.010504e+01
#> 241           257             230 1.559587e+01     1.620196e+01 1.683161e+01
#> 242           219             268 1.059724e+01     1.100850e+01 1.143572e+01
#> 243           213             274 1.000728e+01     1.039524e+01 1.079823e+01
#> 244           251             236 1.462720e+01     1.519589e+01 1.578670e+01
#> 245           181             306 7.467214e+00     7.754318e+00 8.052461e+00
#> 246           218             269 1.049592e+01     1.090318e+01 1.132624e+01
#> 247           256             231 1.542866e+01     1.602831e+01 1.665127e+01
#> 248           245             242 1.373620e+01     1.427035e+01 1.482527e+01
#> 249           236             251 1.252821e+01     1.301529e+01 1.352132e+01
#> 250           226             261 1.134255e+01     1.178317e+01 1.224090e+01
#> 251           244             243 1.359467e+01     1.412333e+01 1.467254e+01
#> 252           235             252 1.240265e+01     1.288483e+01 1.338575e+01
#> 253           255             232 1.526381e+01     1.585710e+01 1.647346e+01
#> 254           212             275 9.913026e+00     1.029725e+01 1.069636e+01
#> 255           180             307 7.401328e+00     7.685817e+00 7.981241e+00
#> 256           197             290 8.623554e+00     8.956616e+00 9.302542e+00
#> 257           217             270 1.039582e+01     1.079912e+01 1.121808e+01
#> 258           196             291 8.545188e+00     8.875136e+00 9.217823e+00
#> 259           259             228 1.593755e+01     1.655680e+01 1.720012e+01
#> 260           248             239 1.417251e+01     1.472359e+01 1.529611e+01
#> 261           211             276 9.819880e+00     1.020041e+01 1.059569e+01
#> 262           225             262 1.123207e+01     1.166834e+01 1.212155e+01
#> 263           234             253 1.227871e+01     1.275604e+01 1.325193e+01
#> 264           210             277 9.727831e+00     1.010472e+01 1.049621e+01
#> 265           224             263 1.112296e+01     1.155493e+01 1.200369e+01
#> 266           223             264 1.101520e+01     1.144293e+01 1.188728e+01
#> 267           161             326 6.269521e+00     6.509331e+00 6.758315e+00
#> 268           233             254 1.215638e+01     1.262892e+01 1.311983e+01
#> 269           195             292 8.467693e+00     8.794559e+00 9.134044e+00
#> 270           232             255 1.203561e+01     1.250342e+01 1.298942e+01
#> 271           231             256 1.191639e+01     1.237953e+01 1.286067e+01
#> 272           179             308 7.336128e+00     7.618031e+00 7.910765e+00
#> 273           254             233 1.510129e+01     1.568830e+01 1.629814e+01
#> 274           250             237 1.447353e+01     1.503628e+01 1.562091e+01
#> 275           266             221 1.721474e+01     1.788302e+01 1.857725e+01
#> 276           216             271 1.029692e+01     1.069632e+01 1.111122e+01
#> 277           247             240 1.402508e+01     1.457045e+01 1.513702e+01
#> 278           194             293 8.391054e+00     8.714874e+00 9.051191e+00
#> 279           139             348 5.194037e+00     5.392460e+00 5.598463e+00
#> 280           193             294 8.315259e+00     8.636066e+00 8.969250e+00
#> 281           253             234 1.494103e+01     1.552186e+01 1.612527e+01
#> 282           178             309 7.271606e+00     7.550949e+00 7.841023e+00
#> 283           192             295 8.240296e+00     8.558122e+00 8.888207e+00
#> 284           209             278 9.636860e+00     1.001014e+01 1.039789e+01
#> 285           215             272 1.019921e+01     1.059476e+01 1.100564e+01
#> 286           243             244 1.345505e+01     1.397828e+01 1.452185e+01
#> 287           191             296 8.166152e+00     8.481031e+00 8.808051e+00
#> 288           208             279 9.546952e+00     9.916671e+00 1.030071e+01
#> 289           114             373 4.197861e+00     4.360517e+00 4.529476e+00
#> 290           207             280 9.458089e+00     9.824285e+00 1.020466e+01
#> 291           222             265 1.090877e+01     1.133231e+01 1.177230e+01
#> 292           160             327 6.215672e+00     6.453376e+00 6.700170e+00
#> 293           159             328 6.162341e+00     6.397960e+00 6.642589e+00
#> 294           177             310 7.207752e+00     7.484563e+00 7.772005e+00
#> 295           158             329 6.109519e+00     6.343077e+00 6.585564e+00
#> 296           176             311 7.144555e+00     7.418862e+00 7.703701e+00
#> 297           190             297 8.092815e+00     8.404778e+00 8.728767e+00
#> 298           138             349 5.150064e+00     5.346838e+00 5.551131e+00
#> 299           189             298 8.020274e+00     8.329353e+00 8.650343e+00
#> 300           221             266 1.080365e+01     1.122305e+01 1.165873e+01
#> 301           175             312 7.082007e+00     7.353836e+00 7.636099e+00
#> 302           174             313 7.020098e+00     7.289476e+00 7.569191e+00
#> 303           206             281 9.370255e+00     9.732968e+00 1.010972e+01
#> 304           205             282 9.283435e+00     9.642704e+00 1.001588e+01
#> 305           188             299 7.948516e+00     8.254743e+00 8.572767e+00
#> 306           137             350 5.106473e+00     5.301617e+00 5.504219e+00
#> 307           173             314 6.958819e+00     7.225773e+00 7.502968e+00
#> 308           204             283 9.197613e+00     9.553477e+00 9.923109e+00
#> 309            80             407 3.105884e+00     3.236765e+00 3.373162e+00
#> 310           187             300 7.877531e+00     8.180936e+00 8.496026e+00
#> 311           186             301 7.807306e+00     8.107920e+00 8.420109e+00
#> 312           214             273 1.010268e+01     1.049440e+01 1.090131e+01
#> 313           136             351 5.063257e+00     5.256791e+00 5.457722e+00
#> 314           203             284 9.112774e+00     9.465270e+00 9.831401e+00
#> 315           172             315 6.898162e+00     7.162717e+00 7.437418e+00
#> 316           202             285 9.028903e+00     9.378068e+00 9.740737e+00
#> 317           171             316 6.838117e+00     7.100299e+00 7.372534e+00
#> 318           113             374 4.161988e+00     4.323440e+00 4.491155e+00
#> 319           112             375 4.126383e+00     4.286648e+00 4.453136e+00
#> 320           135             352 5.020413e+00     5.212354e+00 5.411634e+00
#> 321            79             408 3.077040e+00     3.207268e+00 3.343008e+00
#> 322           111             376 4.091042e+00     4.250135e+00 4.415416e+00
#> 323           170             317 6.778676e+00     7.038511e+00 7.308305e+00
#> 324           169             318 6.719831e+00     6.977343e+00 7.244723e+00
#> 325           168             319 6.661572e+00     6.916787e+00 7.181780e+00
#> 326           201             286 8.945984e+00     9.291857e+00 9.651102e+00
#> 327           134             353 4.977935e+00     5.168302e+00 5.365950e+00
#> 328           110             377 4.055960e+00     4.213899e+00 4.377989e+00
#> 329           157             330 6.057201e+00     6.288720e+00 6.529088e+00
#> 330           230             257 1.179868e+01     1.225721e+01 1.273356e+01
#> 331            78             409 3.048341e+00     3.177932e+00 3.313032e+00
#> 332           156             331 6.005379e+00     6.234881e+00 6.473153e+00
#> 333           200             287 8.864004e+00     9.206621e+00 9.562480e+00
#> 334           109             378 4.021134e+00     4.177936e+00 4.340852e+00
#> 335           185             302 7.737831e+00     8.035684e+00 8.345002e+00
#> 336           184             303 7.669095e+00     7.964217e+00 8.270696e+00
#> 337           199             288 8.782948e+00     9.122345e+00 9.474856e+00
#> 338           155             332 5.954046e+00     6.181553e+00 6.417753e+00
#> 339            77             410 3.019784e+00     3.148753e+00 3.283231e+00
#> 340           154             333 5.903195e+00     6.128729e+00 6.362879e+00
#> 341           133             354 4.935817e+00     5.124630e+00 5.320666e+00
#> 342           220             267 1.069981e+01     1.111512e+01 1.154654e+01
#> 343           153             334 5.852821e+00     6.076402e+00 6.308525e+00
#> 344           167             320 6.603892e+00     6.856834e+00 7.119465e+00
#> 345           198             289 8.702803e+00     9.039015e+00 9.388215e+00
#> 346           108             379 3.986560e+00     4.142240e+00 4.304001e+00
#> 347            76             411 2.991365e+00     3.119729e+00 3.253601e+00
#> 348            75             412 2.963080e+00     3.090854e+00 3.224139e+00
#> 349           152             335 5.802915e+00     6.024567e+00 6.254685e+00
#> 350            74             413 2.934926e+00     3.062126e+00 3.194840e+00
#> 351           151             336 5.753473e+00     5.973215e+00 6.201350e+00
#> 352           132             355 4.894055e+00     5.081332e+00 5.275774e+00
#> 353           107             380 3.952234e+00     4.106810e+00 4.267431e+00
#> 354            73             414 2.906899e+00     3.033542e+00 3.165702e+00
#> 355            72             415 2.878995e+00     3.005097e+00 3.136721e+00
#> 356           131             356 4.852644e+00     5.038403e+00 5.231272e+00
#> 357           106             381 3.918152e+00     4.071640e+00 4.231140e+00
#> 358           166             321 6.546783e+00     6.797477e+00 7.057771e+00
#> 359           165             322 6.490236e+00     6.738706e+00 6.996689e+00
#> 360           164             323 6.434244e+00     6.680514e+00 6.936210e+00
#> 361           150             337 5.704487e+00     5.922341e+00 6.148515e+00
#> 362           149             338 5.655951e+00     5.871938e+00 6.096173e+00
#> 363           183             304 7.601087e+00     7.893508e+00 8.197177e+00
#> 364           130             357 4.811580e+00     4.995838e+00 5.187152e+00
#> 365            71             416 2.851212e+00     2.976787e+00 3.107893e+00
#> 366           105             382 3.884310e+00     4.036727e+00 4.195124e+00
#> 367           148             339 5.607859e+00     5.821999e+00 6.044317e+00
#> 368            70             417 2.823544e+00     2.948609e+00 3.079214e+00
#> 369           147             340 5.560206e+00     5.772519e+00 5.992940e+00
#> 370           163             324 6.378800e+00     6.622893e+00 6.876327e+00
#> 371           104             383 3.850706e+00     4.002067e+00 4.159378e+00
#> 372           129             358 4.770856e+00     4.953632e+00 5.143411e+00
#> 373           128             359 4.730469e+00     4.911782e+00 5.100043e+00
#> 374           103             384 3.817334e+00     3.967656e+00 4.123898e+00
#> 375            69             418 2.795989e+00     2.920560e+00 3.050681e+00
#> 376            68             419 2.768542e+00     2.892635e+00 3.022289e+00
#> 377            67             420 2.741200e+00     2.864830e+00 2.994036e+00
#> 378            66             421 2.713959e+00     2.837142e+00 2.965916e+00
#> 379           102             385 3.784192e+00     3.933492e+00 4.088682e+00
#> 380           101             386 3.751275e+00     3.899569e+00 4.053724e+00
#> 381            65             422 2.686814e+00     2.809566e+00 2.937927e+00
#> 382            64             423 2.659762e+00     2.782099e+00 2.910064e+00
#> 383           146             341 5.512984e+00     5.723492e+00 5.942038e+00
#> 384           145             342 5.466188e+00     5.674910e+00 5.891602e+00
#> 385            63             424 2.632798e+00     2.754736e+00 2.882322e+00
#> 386           127             360 4.690414e+00     4.870281e+00 5.057044e+00
#> 387           100             387 3.718581e+00     3.865884e+00 4.019023e+00
#> 388            62             425 2.605919e+00     2.727474e+00 2.854699e+00
#> 389           162             325 6.323894e+00     6.565835e+00 6.817031e+00
#> 390           126             361 4.650686e+00     4.829124e+00 5.014409e+00
#> 391            99             388 3.686105e+00     3.832435e+00 3.984573e+00
#> 392            61             426 2.579119e+00     2.700307e+00 2.827189e+00
#> 393            98             389 3.653844e+00     3.799216e+00 3.950372e+00
#> 394            97             390 3.621795e+00     3.766225e+00 3.916415e+00
#> 395            60             427 2.552395e+00     2.673231e+00 2.799787e+00
#> 396            96             391 3.589953e+00     3.733458e+00 3.882699e+00
#> 397           125             362 4.611281e+00     4.788309e+00 4.972132e+00
#> 398           124             363 4.572194e+00     4.747829e+00 4.930210e+00
#> 399            59             428 2.525742e+00     2.646242e+00 2.772491e+00
#> 400            58             429 2.499156e+00     2.619336e+00 2.745295e+00
#> 401            57             430 2.472631e+00     2.592507e+00 2.718194e+00
#> 402            56             431 2.446163e+00     2.565750e+00 2.691183e+00
#> 403            55             432 2.419747e+00     2.539061e+00 2.664258e+00
#> 404           123             364 4.533421e+00     4.707680e+00 4.888637e+00
#> 405            54             433 2.393378e+00     2.512435e+00 2.637414e+00
#> 406            53             434 2.367051e+00     2.485866e+00 2.610645e+00
#> 407            52             435 2.340759e+00     2.459348e+00 2.583945e+00
#> 408            95             392 3.558315e+00     3.700911e+00 3.849221e+00
#> 409           144             343 5.419812e+00     5.626769e+00 5.841628e+00
#> 410            94             393 3.526878e+00     3.668581e+00 3.815977e+00
#> 411            51             436 2.314498e+00     2.432877e+00 2.557310e+00
#> 412            50             437 2.288261e+00     2.406445e+00 2.530733e+00
#> 413            49             438 2.262043e+00     2.380048e+00 2.504209e+00
#> 414            48             439 2.235836e+00     2.353678e+00 2.477730e+00
#> 415           143             344 5.373851e+00     5.579062e+00 5.792109e+00
#> 416            47             440 2.209635e+00     2.327329e+00 2.451291e+00
#> 417            93             394 3.495638e+00     3.636464e+00 3.782963e+00
#> 418            92             395 3.464592e+00     3.604557e+00 3.750176e+00
#> 419            91             396 3.433736e+00     3.572856e+00 3.717612e+00
#> 420            90             397 3.403067e+00     3.541358e+00 3.685269e+00
#> 421            46             441 2.183433e+00     2.300994e+00 2.424884e+00
#> 422            45             442 2.157222e+00     2.274666e+00 2.398503e+00
#> 423           142             345 5.328298e+00     5.531783e+00 5.743038e+00
#> 424           122             365 4.494957e+00     4.667858e+00 4.847410e+00
#> 425           121             366 4.456797e+00     4.628358e+00 4.806523e+00
#> 426            89             398 3.372581e+00     3.510059e+00 3.653141e+00
#> 427            44             443 2.130995e+00     2.248337e+00 2.372140e+00
#> 428            43             444 2.104744e+00     2.221999e+00 2.345786e+00
#> 429            42             445 2.078461e+00     2.195644e+00 2.319435e+00
#> 430            88             399 3.342275e+00     3.478956e+00 3.621227e+00
#> 431           120             367 4.418939e+00     4.589176e+00 4.765971e+00
#> 432            87             400 3.312145e+00     3.448046e+00 3.589522e+00
#> 433           119             368 4.381376e+00     4.550307e+00 4.725752e+00
#> 434            41             446 2.052136e+00     2.169264e+00 2.293076e+00
#> 435            40             447 2.025761e+00     2.142848e+00 2.266701e+00
#> 436            86             401 3.282188e+00     3.417324e+00 3.558023e+00
#> 437            39             448 1.999327e+00     2.116387e+00 2.240301e+00
#> 438           141             346 5.283149e+00     5.484927e+00 5.694411e+00
#> 439            38             449 1.972821e+00     2.089870e+00 2.213863e+00
#> 440            37             450 1.946234e+00     2.063287e+00 2.187379e+00
#> 441           140             347 5.238397e+00     5.438488e+00 5.646221e+00
#> 442            36             451 1.919554e+00     2.036625e+00 2.160836e+00
#> 443            85             402 3.252401e+00     3.386788e+00 3.526727e+00
#> 444            35             452 1.892769e+00     2.009872e+00 2.134220e+00
#> 445            34             453 1.865864e+00     1.983014e+00 2.107520e+00
#> 446            33             454 1.838825e+00     1.956037e+00 2.080721e+00
#> 447            32             455 1.811638e+00     1.928925e+00 2.053806e+00
#> 448            31             456 1.784285e+00     1.901662e+00 2.026760e+00
#> 449            30             457 1.756748e+00     1.874227e+00 1.999563e+00
#> 450           118             369 4.344106e+00     4.511748e+00 4.685859e+00
#> 451            29             458 1.729007e+00     1.846603e+00 1.972197e+00
#> 452            28             459 1.701042e+00     1.818767e+00 1.944639e+00
#> 453            84             403 3.222779e+00     3.356433e+00 3.495630e+00
#> 454            27             460 1.672829e+00     1.790695e+00 1.916866e+00
#> 455            26             461 1.644341e+00     1.762361e+00 1.888852e+00
#> 456           117             370 4.307123e+00     4.473493e+00 4.646290e+00
#> 457            25             462 1.615551e+00     1.733737e+00 1.860568e+00
#> 458            24             463 1.586428e+00     1.704789e+00 1.831982e+00
#> 459            23             464 1.556936e+00     1.675484e+00 1.803059e+00
#> 460            22             465 1.527037e+00     1.645781e+00 1.773758e+00
#> 461            21             466 1.496687e+00     1.615635e+00 1.744036e+00
#> 462            20             467 1.465836e+00     1.584995e+00 1.713841e+00
#> 463            19             468 1.434430e+00     1.553805e+00 1.683115e+00
#> 464            18             469 1.402403e+00     1.521999e+00 1.651793e+00
#> 465            17             470 1.369683e+00     1.489500e+00 1.619798e+00
#> 466            16             471 1.336183e+00     1.456220e+00 1.587041e+00
#> 467            15             472 1.301803e+00     1.422055e+00 1.553416e+00
#> 468            83             404 3.193320e+00     3.326257e+00 3.464728e+00
#> 469            14             473 1.266425e+00     1.386883e+00 1.518798e+00
#> 470            13             474 1.229907e+00     1.350555e+00 1.483038e+00
#> 471            12             475 1.192076e+00     1.312891e+00 1.445952e+00
#> 472           116             371 4.270424e+00     4.435539e+00 4.607039e+00
#> 473           115             372 4.234005e+00     4.397882e+00 4.568102e+00
#> 474            11             476 1.152719e+00     1.273670e+00 1.407312e+00
#> 475            10             477 1.111572e+00     1.232612e+00 1.366832e+00
#> 476             9             478 1.068293e+00     1.189358e+00 1.324144e+00
#> 477            82             405 3.164020e+00     3.296256e+00 3.434018e+00
#> 478             8             479 1.022436e+00     1.143437e+00 1.278757e+00
#> 479             7             480 9.733958e-01     1.094204e+00 1.230006e+00
#> 480             6             481 9.203230e-01     1.040753e+00 1.176943e+00
#> 481            81             406 3.134876e+00     3.266427e+00 3.403498e+00
#> 482             5             482 8.619592e-01     9.817321e-01 1.118148e+00
#> 483             4             483 7.963115e-01     9.149839e-01 1.051342e+00
#> 484             3             484 7.198935e-01     8.367029e-01 9.724656e-01
#> 485             2             485 6.255635e-01     7.390033e-01 8.730144e-01
#> 486             1             486 4.935016e-01     5.996204e-01 7.285583e-01
#>     BC_rank    N abundance
#> 1         1    0    368110
#> 2         2    1     70836
#> 3         3    2     44127
#> 4         4    3     32625
#> 5         5    4     25910
#> 6         6    5     21627
#> 7         7    6     18202
#> 8         8    7     15595
#> 9         9    8     13507
#> 10       10    9     12101
#> 11       11   10     10751
#> 12       12   11      9510
#> 13       13   12      8658
#> 14       14   13      7815
#> 15       15   14      7224
#> 16       16   15      6427
#> 17       17   16      6085
#> 18       18   17      5493
#> 19       19   18      5205
#> 20       20   19      4697
#> 21       21   20      4347
#> 22       22   21      3976
#> 23       23   22      3776
#> 24       24   23      3499
#> 25       25   24      3320
#> 26       26   25      3065
#> 27       27   26      2986
#> 28       28   27      2726
#> 29       29   28      2499
#> 30       30   29      2354
#> 31       31   30      2298
#> 32       32   31      2125
#> 33       33   32      1992
#> 34       34   33      1869
#> 35       36   34      1721
#> 36       35   35      1746
#> 37       37   36      1680
#> 38       38   37      1537
#> 39       39   38      1481
#> 40       40   39      1467
#> 41       42   40      1228
#> 42       41   41      1242
#> 43       44   42      1116
#> 44       43   43      1138
#> 45       45   44      1050
#> 46       47   45       995
#> 47       46   46       996
#> 48       48   47       967
#> 49       49   48       919
#> 50       52   49       779
#> 51       51   50       799
#> 52       50   51       848
#> 53       53   52       761
#> 54       54   53       750
#> 55       55   54       666
#> 56       56   55       643
#> 57       57   56       643
#> 58       58   57       611
#> 59       60   58       588
#> 60       59   59       599
#> 61       61   60       563
#> 62       63   61       510
#> 63       64   62       508
#> 64       62   63       534
#> 65       65   64       454
#> 66       68   65       442
#> 67       66   66       450
#> 68       67   67       444
#> 69       69   68       406
#> 70       71   69       386
#> 71       73   70       358
#> 72       70   71       392
#> 73       72   72       368
#> 74       75   73       353
#> 75       76   74       349
#> 76       77   75       311
#> 77       74   76       354
#> 78       78   77       295
#> 79       81   78       271
#> 80       79   79       294
#> 81       82   80       270
#> 82       80   81       278
#> 83       84   82       256
#> 84       83   83       258
#> 85       85   84       242
#> 86       89   85       224
#> 87       88   86       227
#> 88       86   87       235
#> 89       90   88       216
#> 90       87   89       234
#> 91       96   90       190
#> 92       92   91       202
#> 93       93   92       195
#> 94       97   93       184
#> 95      100   94       175
#> 96       91   95       205
#> 97       94   96       194
#> 98       99   97       177
#> 99       95   98       191
#> 100      98   99       182
#> 101     101  100       164
#> 102     102  101       161
#> 103     107  102       133
#> 104     108  103       131
#> 105     103  104       140
#> 106     104  105       139
#> 107     105  106       138
#> 108     106  107       134
#> 109     109  108       129
#> 110     112  109       116
#> 111     110  110       124
#> 112     126  111        93
#> 113     117  112       108
#> 114     113  113       115
#> 115     122  114       102
#> 116     116  115       110
#> 117     114  116       114
#> 118     120  117       105
#> 119     121  118       104
#> 120     119  119       106
#> 121     118  120       107
#> 122     115  121       111
#> 123     111  122       121
#> 124     124  123       101
#> 125     127  124        88
#> 126     128  125        87
#> 127     125  126        94
#> 128     123  127       102
#> 129     134  128        76
#> 130     132  129        83
#> 131     130  130        85
#> 132     129  131        87
#> 133     137  132        71
#> 134     139  133        66
#> 135     143  134        62
#> 136     135  135        76
#> 137     133  136        79
#> 138     142  137        63
#> 139     140  138        64
#> 140     146  139        58
#> 141     136  140        74
#> 142     138  141        71
#> 143     160  142        45
#> 144     154  143        50
#> 145     156  144        49
#> 146     145  145        59
#> 147     157  146        49
#> 148     158  147        48
#> 149     147  148        56
#> 150     144  149        60
#> 151     141  150        64
#> 152     149  151        55
#> 153     152  152        52
#> 154     151  153        53
#> 155     148  154        56
#> 156     161  155        43
#> 157     176  156        35
#> 158     150  157        54
#> 159     165  158        42
#> 160     171  159        38
#> 161     153  160        51
#> 162     182  161        31
#> 163     172  162        38
#> 164     162  163        43
#> 165     170  164        39
#> 166     163  165        43
#> 167     183  166        31
#> 168     166  167        42
#> 169     164  168        43
#> 170     178  169        34
#> 171     155  170        50
#> 172     167  171        42
#> 173     173  172        38
#> 174     189  173        26
#> 175     179  174        33
#> 176     177  175        35
#> 177     193  176        24
#> 178     175  177        37
#> 179     184  178        31
#> 180     159  179        46
#> 181     180  180        32
#> 182     169  181        40
#> 183     215  182        18
#> 184     186  183        30
#> 185     191  184        25
#> 186     190  185        26
#> 187     168  186        42
#> 188     196  187        23
#> 189     210  188        19
#> 190     174  189        38
#> 191     199  190        21
#> 192     188  191        27
#> 193     131  192        84
#> 194     185  193        31
#> 195     205  194        20
#> 196     192  195        25
#> 197     200  196        21
#> 198     181  197        32
#> 199     211  198        19
#> 200     197  199        22
#> 201     201  200        21
#> 202     217  201        17
#> 203     194  202        24
#> 204     206  203        20
#> 205     212  204        19
#> 206     202  205        21
#> 207     187  206        30
#> 208     195  207        24
#> 209     207  208        20
#> 210     203  209        21
#> 211     198  210        22
#> 212     208  211        20
#> 213     222  212        16
#> 214     204  213        21
#> 215     245  214        11
#> 216     223  215        16
#> 217     218  216        17
#> 218     219  217        17
#> 219     246  218        11
#> 220     247  219        11
#> 221     248  220        11
#> 222     235  221        14
#> 223     224  222        16
#> 224     220  223        17
#> 225     225  224        16
#> 226     226  225        16
#> 227     227  226        16
#> 228     209  227        20
#> 229     258  228        10
#> 230     213  229        19
#> 231     238  230        13
#> 232     305  231         6
#> 233     249  232        11
#> 234     216  233        18
#> 235     259  234        10
#> 236     229  235        15
#> 237     241  236        12
#> 238     250  237        11
#> 239     260  238        10
#> 240     214  239        19
#> 241     230  240        15
#> 242     268  241         9
#> 243     274  242         8
#> 244     236  243        14
#> 245     306  244         6
#> 246     269  245         9
#> 247     231  246        15
#> 248     242  247        12
#> 249     251  248        11
#> 250     261  249        10
#> 251     243  250        12
#> 252     252  251        11
#> 253     232  252        15
#> 254     275  253         8
#> 255     307  254         6
#> 256     290  255         7
#> 257     270  256         9
#> 258     291  257         7
#> 259     228  258        16
#> 260     239  259        13
#> 261     276  260         8
#> 262     262  261        10
#> 263     253  262        11
#> 264     277  263         8
#> 265     263  264        10
#> 266     264  265        10
#> 267     326  266         5
#> 268     254  267        11
#> 269     292  268         7
#> 270     255  269        11
#> 271     256  270        11
#> 272     308  271         6
#> 273     233  272        15
#> 274     237  273        14
#> 275     221  274        17
#> 276     271  275         9
#> 277     240  276        13
#> 278     293  277         7
#> 279     348  278         4
#> 280     294  279         7
#> 281     234  280        15
#> 282     309  281         6
#> 283     295  282         7
#> 284     278  283         8
#> 285     272  284         9
#> 286     244  285        12
#> 287     296  286         7
#> 288     279  287         8
#> 289     373  288         3
#> 290     280  289         8
#> 291     265  290        10
#> 292     327  291         5
#> 293     328  292         5
#> 294     310  293         6
#> 295     329  294         5
#> 296     311  295         6
#> 297     297  296         7
#> 298     349  297         4
#> 299     298  298         7
#> 300     266  299        10
#> 301     312  300         6
#> 302     313  301         6
#> 303     281  302         8
#> 304     282  303         8
#> 305     299  305         7
#> 306     350  306         4
#> 307     314  307         6
#> 308     283  308         8
#> 309     407  309         2
#> 310     300  310         7
#> 311     301  311         7
#> 312     273  312         9
#> 313     351  313         4
#> 314     284  314         8
#> 315     315  315         6
#> 316     285  316         8
#> 317     316  317         6
#> 318     374  318         3
#> 319     375  319         3
#> 320     352  320         4
#> 321     408  321         2
#> 322     376  322         3
#> 323     317  323         6
#> 324     318  324         6
#> 325     319  326         6
#> 326     286  327         8
#> 327     353  328         4
#> 328     377  329         3
#> 329     330  330         5
#> 330     257  331        11
#> 331     409  332         2
#> 332     331  333         5
#> 333     287  335         8
#> 334     378  336         3
#> 335     302  337         7
#> 336     303  338         7
#> 337     288  339         8
#> 338     332  340         5
#> 339     410  341         2
#> 340     333  342         5
#> 341     354  343         4
#> 342     267  344        10
#> 343     334  345         5
#> 344     320  346         6
#> 345     289  347         8
#> 346     379  348         3
#> 347     411  349         2
#> 348     412  350         2
#> 349     335  351         5
#> 350     413  353         2
#> 351     336  356         5
#> 352     355  358         4
#> 353     380  359         3
#> 354     414  360         2
#> 355     415  361         2
#> 356     356  362         4
#> 357     381  363         3
#> 358     321  364         6
#> 359     322  365         6
#> 360     323  366         6
#> 361     337  367         5
#> 362     338  369         5
#> 363     304  371         7
#> 364     357  373         4
#> 365     416  374         2
#> 366     382  375         3
#> 367     339  376         5
#> 368     417  377         2
#> 369     340  378         5
#> 370     324  379         6
#> 371     383  380         3
#> 372     358  381         4
#> 373     359  382         4
#> 374     384  383         3
#> 375     418  384         2
#> 376     419  386         2
#> 377     420  387         2
#> 378     421  388         2
#> 379     385  389         3
#> 380     386  390         3
#> 381     422  391         2
#> 382     423  392         2
#> 383     341  393         5
#> 384     342  394         5
#> 385     424  395         2
#> 386     360  396         4
#> 387     387  397         3
#> 388     425  398         2
#> 389     325  399         6
#> 390     361  401         4
#> 391     388  405         3
#> 392     426  406         2
#> 393     389  407         3
#> 394     390  408         3
#> 395     427  409         2
#> 396     391  412         3
#> 397     362  413         4
#> 398     363  414         4
#> 399     428  418         2
#> 400     429  419         2
#> 401     430  420         2
#> 402     431  421         2
#> 403     432  422         2
#> 404     364  424         4
#> 405     433  425         2
#> 406     434  429         2
#> 407     435  431         2
#> 408     392  433         3
#> 409     343  435         5
#> 410     393  437         3
#> 411     436  439         2
#> 412     437  440         2
#> 413     438  441         2
#> 414     439  442         2
#> 415     344  443         5
#> 416     440  446         2
#> 417     394  448         3
#> 418     395  449         3
#> 419     396  450         3
#> 420     397  451         3
#> 421     441  452         2
#> 422     442  453         2
#> 423     345  454         5
#> 424     365  456         4
#> 425     366  460         4
#> 426     398  464         3
#> 427     443  465         2
#> 428     444  466         2
#> 429     445  467         2
#> 430     399  470         3
#> 431     367  471         4
#> 432     400  474         3
#> 433     368  475         4
#> 434     446  479         2
#> 435     447  482         2
#> 436     401  484         3
#> 437     448  489         2
#> 438     346  491         5
#> 439     449  492         2
#> 440     450  493         2
#> 441     347  496         5
#> 442     451  499         2
#> 443     402  500         3
#> 444     452  502         2
#> 445     453  503         2
#> 446     454  510         2
#> 447     455  512         2
#> 448     456  518         2
#> 449     457  522         2
#> 450     369  524         4
#> 451     458  525         2
#> 452     459  527         2
#> 453     403  528         3
#> 454     460  529         2
#> 455     461  530         2
#> 456     370  531         4
#> 457     462  532         2
#> 458     463  533         2
#> 459     464  535         2
#> 460     465  543         2
#> 461     466  546         2
#> 462     467  547         2
#> 463     468  549         2
#> 464     469  557         2
#> 465     470  567         2
#> 466     471  581         2
#> 467     472  582         2
#> 468     404  594         3
#> 469     473  599         2
#> 470     474  601         2
#> 471     475  615         2
#> 472     371  622         4
#> 473     372  632         4
#> 474     476  634         2
#> 475     477  635         2
#> 476     478  660         2
#> 477     405  695         3
#> 478     479  697         2
#> 479     480  723         2
#> 480     481  770         2
#> 481     406  853         3
#> 482     482  903         2
#> 483     483  954         2
#> 484     484  973         2
#> 485     485  988         2
#> 486     486 1230         2
#> 
#> [[2]]
#>            A            a            b 
#> 5.482970e+05 2.218812e+00 2.949379e-01 
#> 
#> [[3]]
#>                    2.5 %       97.5 %
#> (Intercept) 3.682324e+05 8.164128e+05
#> log_den     2.260682e+00 2.176941e+00
#> log_num     2.530679e-01 3.368079e-01
#> 
#> [[4]]
#> 
#> Call:
#> stats::lm(formula = log_abundance ~ log_den + log_num)
#> 
#> Coefficients:
#> (Intercept)      log_den      log_num  
#>     13.2146      -2.2188       0.2949  
#> 
#> 
#> [[5]]
#> 
#> Call:
#> stats::lm(formula = log_abundance ~ log_den + log_num)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -2.33243 -0.18667 -0.06665  0.27185  1.20461 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept) 13.21457    0.20261   65.22   <2e-16 ***
#> log_den     -2.21881    0.02131 -104.12   <2e-16 ***
#> log_num      0.29494    0.02131   13.84   <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 0.3417 on 483 degrees of freedom
#> Multiple R-squared:  0.9796,	Adjusted R-squared:  0.9796 
#> F-statistic: 1.162e+04 on 2 and 483 DF,  p-value: < 2.2e-16
#> 
#> 
#> [[6]]
#> function (rank) 
#> {
#>     params["A"] * (max(t_frame[, "BC_rank"]) + 1 - rank)^params["b"]/(rank^params["a"])
#> }
#> <bytecode: 0x558ca1d791f0>
#> <environment: 0x558cb2de2778>
#>